requirements.txt
# Python 3.12 and above required (3.12 preferred for stability and reliability)

torch
torchvision
torchaudio
numpy
matplotlib
seaborn
scikit-learn
Pillow
tqdm
jupyter
ipykernel
kaggle
scipy
pyyaml
textual
textual-plotext
plotext